change port identifiers to use uint16_t

+++ src/apps/relay/tls_listener.h
@@ -41,6 +41,8 @@
 
 #include "ns_ioalib_impl.h"
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -52,7 +54,7 @@ typedef struct tls_listener_relay_server_info tls_listener_relay_server_type;
 
 ///////////////////////////////////////////
 
-tls_listener_relay_server_type *create_tls_listener_server(const char *ifname, const char *local_address, int port,
+tls_listener_relay_server_type *create_tls_listener_server(const char *ifname, const char *local_address, uint16_t port,
                                                            int verbose, ioa_engine_handle e,
                                                            ioa_engine_new_connection_event_handler send_socket,
                                                            struct relay_server *relay_server);
